+++ src/main/java/wooteco/member/controller/advice/MemberControllerAdvice.java
@@ -5,7 +5,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
-import wooteco.exception.EmailNotFoundException;
+import wooteco.member.exception.EmailNotFoundException;
 
 import java.io.InvalidClassException;
