+++ src/main/java/wooteco/member/exception/EmailNotFoundException.java
@@ -1,4 +1,4 @@
-package wooteco.exception;
+package wooteco.member.exception;
 
 public class EmailNotFoundException extends RuntimeException {
     public EmailNotFoundException(String s) {
