+++ src/main/java/wooteco/member/exception/InvalidPasswordException.java
@@ -1,4 +1,4 @@
-package wooteco.exception;
+package wooteco.member.exception;
 
 public class InvalidPasswordException extends RuntimeException {
     public InvalidPasswordException(String s) {
