+++ src/main/java/wooteco/member/service/AuthService.java
@@ -4,8 +4,8 @@ import io.jsonwebtoken.JwtException;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
-import wooteco.exception.EmailNotFoundException;
-import wooteco.exception.InvalidPasswordException;
+import wooteco.member.exception.EmailNotFoundException;
+import wooteco.member.exception.InvalidPasswordException;
 import wooteco.member.controller.dto.request.ApprovedMemberRequest;
 import wooteco.member.controller.dto.request.SignInRequest;
 import wooteco.member.controller.dto.response.SignInResponse;
